pg_user_function/requirements.txt
azure-functions
psycopg2-binary
typing

